Move database tests to a shared location and create tests package including them.
