+++ test/package.json
@@ -1,6 +1,14 @@
 {
-  "name": "test",
-  "private": true,
+  "name": "@malloydata/malloy-tests",
+  "version": "0.0.63",
+  "license": "MIT",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "homepage": "https://github.com/malloydata/malloy#readme",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/malloydata/malloy"
+  },
   "scripts": {
     "lint": "eslint '**/*.ts{,x}'",
     "lint-fix": "eslint '**/*.ts{,x}' --fix",
@@ -11,11 +19,11 @@
   },
   "dependencies": {
     "@jest/globals": "^29.4.3",
-    "@malloydata/db-bigquery": "^0.0.63",
-    "@malloydata/db-duckdb": "^0.0.63",
-    "@malloydata/db-postgres": "^0.0.63",
-    "@malloydata/malloy": "^0.0.63",
-    "@malloydata/render": "^0.0.63",
+    "@malloydata/db-bigquery": "^0.0.62",
+    "@malloydata/db-duckdb": "^0.0.62",
+    "@malloydata/db-postgres": "^0.0.62",
+    "@malloydata/malloy": "^0.0.62",
+    "@malloydata/render": "^0.0.62",
     "jsdom": "^22.1.0",
     "luxon": "^1.26.4",
     "madge": "^6.0.0"
@@ -23,6 +31,5 @@
   "devDependencies": {
     "@types/jsdom": "^21.1.1",
     "@types/luxon": "^1.26.4"
-  },
-  "version": "0.0.63"
+  }
 }
